Исправлено открытие torrent файлов

+++ app/src/main/java/su/css3/openwithelementum/bencode/BencodeInt.java
@@ -4,24 +4,24 @@ import java.io.IOException;
 import java.io.OutputStream;
 
 public class BencodeInt implements BencodeObject {
-    private int value;
+    private long value; // Yes, BencodeInt is long
 
-    public BencodeInt(int value) {
+    public BencodeInt(long value) {
         this.value = value;
     }
 
-    public int getInt() {
+    public long getLong() {
         return this.value;
     }
 
     public void write(OutputStream out) throws IOException {
         out.write('i');
-        out.write(Integer.toString(this.value).getBytes());
+        out.write(Long.toString(this.value).getBytes());
         out.write('e');
     }
 
     @Override
     public String toString() {
-        return Integer.toString(this.value);
+        return Long.toString(this.value);
     }
 }
